Rewrite and reformat the neural network implementation
The new NN class is now a simple 2-layer network; this will be slightly
extended later on. Parameter initialization and the optimizer were
changed, and a (temporary) simple test/demo was added.

It turns out the network works better with one-hot coded outputs.
All in all, this should be much cleaner.